重命名parentId->dictTypeId和sort->sortOrder
重构字典和菜单模型，改用dictTypeId代替parentId，并将排序字段统一为前端和后端的sortOrder。变更涉及Vue组件、TS类型、API调用、Java DTO/实体/服务/控制器、MyBatis XML 映射器和缓存键。还增加了 FileUpload 的最大 ChunkSize，更新了生产的临时/存储路径为 /tmp，并将 SysDialogue.sortOrder 类型改为整数。相应地更新了控制器参数名称和签名。

+++ imba-radish/src/main/java/com/soft/sys/entity/SysDialogue.java
@@ -18,7 +18,7 @@ public class SysDialogue extends BaseEntity {
      * 排序字段
      */
     @TableField(value = "sort_order")
-    private Long sortOrder;
+    private Integer sortOrder;
 
     /**
      * 标题
